feat: ElevenLabs TTS + Análisis de sentimientos + Correcciones
- Integración ElevenLabs para respuestas con audio en WhatsApp
- Análisis de sentimientos para respuestas más humanas
- Corregir nombre del agente en respuestas (usa nombre_custom)
- Fix SQL: CURRENT_DATE - $1::integer en métricas
- Dockerfile: incluir node_modules para PDF/Word/Excel parsing
- Nuevo endpoint /api/elevenlabs/voces para listar voces
- UI en Agentes para configurar voz y respuestas de audio

+++ app/api/metricas/api-usage/route.ts
@@ -32,7 +32,7 @@ export async function GET(request: NextRequest) {
               m.vision_costo_usd + m.elevenlabs_costo_usd + m.twilio_costo_usd) as total_costo
          FROM metricas_api m
          JOIN clientes c ON m.cliente_id = c.id
-         WHERE m.fecha >= CURRENT_DATE - $1
+         WHERE m.fecha >= CURRENT_DATE - $1::integer
          GROUP BY c.id, c.nombre_empresa
          ORDER BY total_costo DESC
          LIMIT 10`,
